implemented edit button and its functionalities

+++ virtualTourist/ViewController/MapViewController.swift
@@ -17,12 +17,9 @@ class MapViewController: UIViewController, MKMapViewDelegate, UIGestureRecognize
     @IBOutlet weak var tapPinsDeleteLabel: UILabel!
     @IBOutlet weak var editButtonOutlet: UIBarButtonItem!
     var mapAnnotationDictionary: Dictionary = [MKPointAnnotation: Map]()
+    var isOnEditMode = false
     
-    @IBAction func editPinsButton(_ sender: Any) {
-        tapPinsDeleteLabel.isHidden = false
-        editButtonOutlet.title = "Done"
-        
-    }
+   
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -62,33 +59,37 @@ class MapViewController: UIViewController, MKMapViewDelegate, UIGestureRecognize
         }
         print("MARCELA DICTIONARY: \(mapAnnotationDictionary)")
         
+        
         let uilgr = UILongPressGestureRecognizer(target: self, action: #selector(addPin(gestureRecognizer:)))
         uilgr.minimumPressDuration = 1.0
         uilgr.delegate = self
         mapView.addGestureRecognizer(uilgr)
+        
     }
     
     //MARK: UIGestureRecognizerDelegate methods
     
     @objc func addPin(gestureRecognizer:UIGestureRecognizer){
         if gestureRecognizer.state == .ended{
-            //obtaining new coordinates
-            let touchPoint = gestureRecognizer.location(in: mapView)
-            let newCoordinates = mapView.convert(touchPoint, toCoordinateFrom: mapView)
-            
-            //creating annotations and adding to map
-            let annotation = MKPointAnnotation()
-            annotation.title = "title"
-            annotation.coordinate = newCoordinates
-            mapView.addAnnotation(annotation)
-            
-            //            init managedObject and add annotations to DataBase
-            let mapLocation = Map(context: DataBaseController.persistentContainer
-                .viewContext)
-            mapLocation.latitude = newCoordinates.latitude
-            mapLocation.longitude = newCoordinates.longitude
-            
-            DataBaseController.saveContext()
+            if !isOnEditMode{
+                //obtaining new coordinates
+                let touchPoint = gestureRecognizer.location(in: mapView)
+                let newCoordinates = mapView.convert(touchPoint, toCoordinateFrom: mapView)
+                
+                //creating annotations and adding to map
+                let annotation = MKPointAnnotation()
+                annotation.title = "title"
+                annotation.coordinate = newCoordinates
+                mapView.addAnnotation(annotation)
+                
+                //            init managedObject and add annotations to DataBase
+                let mapLocation = Map(context: DataBaseController.persistentContainer
+                    .viewContext)
+                mapLocation.latitude = newCoordinates.latitude
+                mapLocation.longitude = newCoordinates.longitude
+                
+                DataBaseController.saveContext()
+            }
         }
     }
     
@@ -110,18 +111,37 @@ class MapViewController: UIViewController, MKMapViewDelegate, UIGestureRecognize
         return pinView
     }
     
-    func mapView(_ mapView: MKMapView, didSelect view: MKAnnotationView) {
-        print("MARCELA: ENTRA NA FUNCAO QUANDO SELECIONO PIN")
-        if let annotation = view.annotation{
-            mapView.removeAnnotation(annotation)
-            if let mapForAnnotation = mapAnnotationDictionary[annotation as! MKPointAnnotation]{
-            DataBaseController.persistentContainer.viewContext.delete(mapForAnnotation)
-            DataBaseController.saveContext()
-            mapAnnotationDictionary[annotation as! MKPointAnnotation] = nil
-            }
-            
+    @IBAction func editPinsButton(_ sender: Any) {
+        isOnEditMode = !isOnEditMode
+        print("MARCELA: \(isOnEditMode)")
+        if isOnEditMode{
+            tapPinsDeleteLabel.isHidden = false
+            editButtonOutlet.title = "Done"
+        }else{
+            tapPinsDeleteLabel.isHidden = true
+            editButtonOutlet.title = "Edit"
         }
     }
+    //essa funcao só será chamada se o edit button for chamado
+    
+    func mapView(_ mapView: MKMapView, didSelect view: MKAnnotationView) {
+        if isOnEditMode{
+            //delete annotation
+            if let annotation = view.annotation{
+                mapView.removeAnnotation(annotation)
+                if let mapForAnnotation = mapAnnotationDictionary[annotation as! MKPointAnnotation]{
+                    DataBaseController.persistentContainer.viewContext.delete(mapForAnnotation)
+                    DataBaseController.saveContext()
+                    mapAnnotationDictionary[annotation as! MKPointAnnotation] = nil
+                }
+                
+            }
+        }else{
+            //faz o request pro flicker e troca a VC
+        }
+    }
+    
+    //caso o pin tenha sido pressionado chama a funcao
     
     
     func mapView(_ mapView: MKMapView, regionDidChangeAnimated animated: Bool) {
